Created and implemented a CircumfrentialProgressBarTimer
This GraphicalTimer displays the time remaining as a portion of a
circular arc proportional to the amount of time remaining out of the set
target time.

Next steps are creating and running unit tests to ensure that the class
can, indeed, function properly and to optimize the rendering code as
best I can to prevent excessive memory usage.