Began Implementing PreferenceSelector
Began the implementation of the PreferenceSelector dialog.

The dialog is to show a list of all available preference options in a
vertical, scrollable list, the current settings for all of the
preferences supported, and allow the user to change their preferences as
needed.

At present, I am implementing the PreferenceListItem which will
represent an entry in the vertically scrolling list of preferences. The
goal is to have it be split down the middle, with the left side naming
the preference and the right allowing for the changing of the
preferences.

The class will also allow for the direct pulling of the preference's
values.

NOTE:  TODOs are tasks that need to be planned and don't have the infrastructure for completing yet.